Add a ClassMap implementation based off shapeless.HMap
Implementation notes: We restrict the key-value pairs in the
`ClassMap` by requiring an implicit value of type `Bounds[U, L]#λ[K, V]`,
where `U` and `L` are upper and lower bounds, respectively,
for the types representable in the map. Since this is a sealed trait,
it only exists when we `materialize` it, which will only typecheck
if `K =:= Class[V]` and `V >: L <: U`. Thus is type safety maintained.

Note that since there is only ever one value of type `Bounds[U, L]#λ[K, V]`,
no matter what `U`, `L`, `K`, and `V` may be, nonsense may occur if
these types are used for anything other than constraining the contents
of a `ClassMap`. Just don't do that.